Move ATTRS to new file to fix circ dep issue
Implements this pattern across
