+++ server/src/packages/packages.attrs.ts
@@ -1,0 +1,11 @@
+export const PACKAGE_ATTRS = [
+  'statuscode',
+  'statecode',
+  'dcp_packagetype',
+  'dcp_visibility',
+  'dcp_packageversion',
+  '_dcp_rwcdsform_value',
+  '_dcp_pasform_value',
+  'dcp_packageid',
+  'dcp_name',
+];
