+++ server/src/packages/pas-form/applicants/applicant-representative.attrs.ts
@@ -1,0 +1,11 @@
+export const APPLICANT_REPRESENTATIVE_ATTRS = [
+  'dcp_firstname',
+  'dcp_lastname',
+  'dcp_organization',
+  'dcp_email',
+  'dcp_address',
+  'dcp_city',
+  'dcp_state',
+  'dcp_zipcode',
+  'dcp_phone',
+];
