+++ server/src/packages/pas-form/applicants/applicants.attrs.ts
@@ -1,0 +1,6 @@
+import { APPLICANT_REPRESENTATIVE_ATTRS } from "./applicant-representative.attrs";
+
+export const APPLICANT_ATTRS = [
+  ...APPLICANT_REPRESENTATIVE_ATTRS,
+  'dcp_type',
+];
