+++ server/src/packages/pas-form/applicants/applicants.controller.ts
@@ -14,6 +14,8 @@ import { CrmService } from '../../../crm/crm.service';
 import { JsonApiSerializeInterceptor } from '../../../json-api-serialize.interceptor';
 import { AuthenticateGuard } from '../../../authenticate.guard';
 import { JsonApiDeserializePipe } from '../../../json-api-deserialize.pipe';
+import { APPLICANT_REPRESENTATIVE_ATTRS } from './applicant-representative.attrs';
+import { APPLICANT_ATTRS } from './applicants.attrs';
 
 /**
 * CRM has two entities for applicants (dcp_applicantinformation and dcp_applicantrepresentativeinformation)
@@ -25,23 +27,6 @@ import { JsonApiDeserializePipe } from '../../../json-api-deserialize.pipe';
 * to make sure there are unique IDs, and so we are able to delete the correct CRM record in our
 * delete method in this controller (which doesn't have access to body.target_entity!)
 **/
-
-export const APPLICANT_REPRESENTATIVE_ATTRS = [
-  'dcp_firstname',
-  'dcp_lastname',
-  'dcp_organization',
-  'dcp_email',
-  'dcp_address',
-  'dcp_city',
-  'dcp_state',
-  'dcp_zipcode',
-  'dcp_phone',
-];
-
-export const APPLICANT_ATTRS = [
-  ...APPLICANT_REPRESENTATIVE_ATTRS,
-  'dcp_type',
-];
 
 @UseInterceptors(
   new JsonApiSerializeInterceptor('applicants', {
