+++ server/src/packages/pas-form/bbls/bbls.attrs.ts
@@ -1,0 +1,8 @@
+export const BBL_ATTRS = [
+  'dcp_partiallot',
+  'dcp_developmentsite',
+  'dcp_bblnumber',
+  'dcp_userinputborough',
+  'dcp_userinputblock',
+  'dcp_userinputlot',
+];
