+++ server/src/packages/pas-form/pas-form.attrs.ts
@@ -108,3 +108,40 @@ export const PAS_FORM_ATTRS = [
   // Internals
   '_dcp_projectname_value',
 ];
+
+export const PAS_FORM_PROJECTADDRESS_ATTRS = [
+  'dcp_validatedpostalcode',
+  'dcp_projectaddressid',
+  'modifiedon',
+  'dcp_dmsourceid',
+  'dcp_name',
+  'dcp_validatedxcoordinate',
+  'overriddencreatedon',
+  'dcp_validatedccd',
+  'createdon',
+  'dcp_userinputborough',
+  'dcp_addressvalidated',
+  'dcp_userinputaddressnumber',
+  'dcp_validatedstreet',
+  'dcp_responsewarning',
+  'dcp_userinputunit',
+  'versionnumber',
+  'dcp_migratedlastupdateddate',
+  'statuscode',
+  'dcp_validatedycoordinate',
+  'dcp_validatedborough',
+  'dcp_responseerror',
+  'dcp_validatedcd',
+  'timezoneruleversionnumber',
+  'dcp_validatedzm',
+  'dcp_validatedaddressnumber',
+  'importsequencenumber',
+  'dcp_dcp_validatedbintext',
+  'utcconversiontimezonecode',
+  'dcp_userinputstreet',
+  'dcp_validatedstreetcode',
+  'dcp_concatenatedaddressvalidated',
+  'dcp_validatedaddressoverride',
+  'dcp_addressvalidateddate',
+  'statecode',
+];
