+++ server/src/projects/projects.attrs.ts
@@ -1,0 +1,10 @@
+export const PROJECT_ATTRS = [
+  'dcp_projectname',
+  'dcp_name',
+  'dcp_borough',
+  'statecode',
+  'statuscode',
+  'dcp_visibility',
+  '_dcp_applicant_customer_value',
+  'dcp_dcp_project_dcp_projectapplicant_Project',
+];
